refactor: moved out base class for issues
